rename 'changelog.py' to 'recovery.py' since it's a bit more encompassing
