+++ quixstreams/state/__init__.py
@@ -1,3 +1,3 @@
 from .manager import *
+from .recovery import *
 from .types import *
-from .changelog import ChangelogManager
